Don't try to use bundler to load rubocop
Using `bundle exec` solves one problem but creates others.

It solves the 'wrong version' of rubocop issue (ie you have a newer version but the repo expects to use an older version)

But, it means you can't get rubocop highlighting if you've not run `bundle` to install all gems - even though you may already have rubocop installed.

It also adds 0.5 sec to the run time of this frequently run command - this delay will increase with the number and size of gems being bundled.

If this choice proves to problematic it can be revisited later.